echo norm OK
